requirements.txt
python-dotenv
replicate
requests